Fix type mismatch error when building on QT app

+++ include/vlc/libvlc_media.h
@@ -344,7 +344,7 @@ typedef int (*libvlc_media_open_cb)(void *opaque, void **datap,
  * \note If no data is immediately available, then the callback should sleep.
  * \warning The application is responsible for avoiding deadlock situations.
  */
-typedef ssize_t (*libvlc_media_read_cb)(void *opaque, unsigned char *buf,
+typedef int (*libvlc_media_read_cb)(void *opaque, unsigned char *buf,
                                         size_t len);
 
 /**
